Fix GroupBasedEncoder::mergeSequenceParams

+++ MivBitstream/src/IvSequenceParams.cpp
@@ -171,7 +171,7 @@ IvSequenceParams::IvSequenceParams(const SizeVector &atlasSizes) {
       .ptl_profile_reconstruction_idc(PtlProfileReconstructionIdc::MIV_Main);
 
   VERIFY_MIVBITSTREAM(!atlasSizes.empty());
-  vps.vps_atlas_count_minus1(atlasSizes.size() - 1);
+  vps.vps_atlas_count_minus1(uint8_t(atlasSizes.size() - 1));
 
   for (size_t atlasId = 0; atlasId < atlasSizes.size(); ++atlasId) {
     const auto a = uint8_t(atlasId);
